+++ packages/nestjs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @prismakit/nestjs
 
+## 4.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @prismakit/core@4.0.2
+  - @prismakit/redis@4.0.2
+
 ## 4.0.1
 
 ### Patch Changes