+++ packages/opentelemetry/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @prismakit/opentelemetry
 
+## 4.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @prismakit/core@4.0.2
+
 ## 4.0.1
 
 ### Patch Changes